+++ packages/mason/lib/src/bricks_json.dart
@@ -1,18 +1,46 @@
 import 'dart:convert';
 
+import 'package:checked_yaml/checked_yaml.dart';
 import 'package:crypto/crypto.dart';
+import 'package:http/http.dart' as http;
 import 'package:mason/mason.dart';
 import 'package:mason/src/git.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
+import 'package:pub_semver/pub_semver.dart';
 import 'package:universal_io/io.dart';
 
-/// {@template write_brick_exception}
-/// Thrown when an error occurs while writing a brick to cache.
+/// {@template brick_resolve_version_exception}
+/// Thrown when an error occurs while resolving a brick version.
 /// {@endtemplate}
-class WriteBrickException extends MasonException {
-  /// {@macro write_brick_exception}
-  const WriteBrickException(String message) : super(message);
+class BrickResolveVersionException extends MasonException {
+  /// {@macro brick_resolve_version_exception}
+  const BrickResolveVersionException(String message) : super(message);
+}
+
+/// {@template mason_yaml_name_mismatch}
+/// Thrown when a brick's name in `mason.yaml` does not match
+/// the name in `brick.yaml`.
+/// {@endtemplate}
+class MasonYamlNameMismatch extends MasonException {
+  /// {@macro mason_yaml_name_mismatch}
+  MasonYamlNameMismatch(String message) : super(message);
+}
+
+/// {@template brick_unsatisfied_version_constraint}
+/// Thrown when a brick version constraint could not be satisfied.
+/// {@endtemplate}
+class BrickUnsatisfiedVersionConstraint extends MasonException {
+  /// {@macro brick_unsatisfied_version_constraint}
+  BrickUnsatisfiedVersionConstraint(String message) : super(message);
+}
+
+/// {@template malformed_bricks_json}
+/// Thrown when a brick from mason.yaml cannot be found in bricks.json.
+/// {@endtemplate}
+class MalformedBricksJson extends MasonException {
+  /// {@macro malformed_bricks_json}
+  const MalformedBricksJson(String message) : super(message);
 }
 
 /// {@template bricks_json}
@@ -51,6 +79,9 @@ class BricksJson {
   /// Associated bricks.json file
   late File _bricksJsonFile;
 
+  /// Read-only cache of brick,location pairs.
+  Map<String, String> get cache => _cache;
+
   /// Removes all key/value pairs from the cache.
   void clear() {
     _cache.clear();
@@ -64,9 +95,15 @@ class BricksJson {
     if (!bricksJson.existsSync()) return <String, String>{};
     final content = bricksJson.readAsStringSync();
     if (content.isEmpty) return <String, String>{};
-    return Map.castFrom<dynamic, dynamic, String, String>(
-      json.decode(content) as Map,
-    );
+    try {
+      return Map.castFrom<dynamic, dynamic, String, String>(
+        json.decode(content) as Map,
+      );
+    } catch (error) {
+      throw MalformedBricksJson(
+        'Malformed bricks.json at ${bricksJson.path}\n$error',
+      );
+    }
   }
 
   /// Encodes entire cache contents.
@@ -87,22 +124,24 @@ class BricksJson {
   /// Returns the cache key for the given [brick].
   /// Return `null` if the [brick] does not have a valid path.
   String? _getKey(Brick brick) {
-    if (brick.path != null) {
-      final path = brick.path!;
-      final name = p.basenameWithoutExtension(path);
-      final hash = sha256.convert(utf8.encode(path));
+    final name = brick.name;
+    final location = brick.location;
+    if (location.path != null) {
+      final hash = sha256.convert(utf8.encode(location.path!));
       final key = '${name}_$hash';
       return key;
     }
-    if (brick.git != null) {
+    if (location.git != null) {
       final path =
-          p.join(brick.git!.url, brick.git!.path).replaceAll(r'\', '/');
-      final name = p.basenameWithoutExtension(path);
+          p.join(location.git!.url, location.git!.path).replaceAll(r'\', '/');
       final hash = sha256.convert(utf8.encode(path));
-      final key = brick.git!.ref != null
-          ? '${name}_${brick.git!.ref}_$hash'
+      final key = location.git!.ref != null
+          ? '${name}_${location.git!.ref}_$hash'
           : '${name}_$hash';
       return key;
+    }
+    if (location.version != null) {
+      return '${name}_${location.version}';
     }
     return null;
   }
@@ -114,41 +153,53 @@ class BricksJson {
   /// Caches brick if necessary and updates `bricks.json`.
   /// Returns the local path to the brick.
   Future<String> add(Brick brick) async {
-    final key = _getKey(brick);
-    if (key == null) {
-      throw const WriteBrickException(
-        'Brick must contain either a path or a git url',
+    if (brick.location.path != null) {
+      return _addLocalBrick(brick);
+    }
+
+    if (brick.location.git != null) {
+      return _addRemoteBrickFromGit(brick);
+    }
+
+    return _addRemoteBrickFromRegistry(brick);
+  }
+
+  /// Writes local brick at using path to cache
+  /// and returns the local path to the brick.
+  String _addLocalBrick(Brick brick) {
+    final path = brick.location.path!;
+    final brickYaml = File(p.join(path, BrickYaml.file));
+
+    if (!brickYaml.existsSync()) {
+      throw BrickNotFoundException(p.canonicalize(path));
+    }
+
+    final yaml = checkedYamlDecode(
+      brickYaml.readAsStringSync(),
+      (m) => BrickYaml.fromJson(m!),
+    );
+
+    if (yaml.name != brick.name) {
+      throw MasonYamlNameMismatch(
+        'Brick name "${brick.name}" '
+        'doesn\'t match provided name "${yaml.name}" in ${MasonYaml.file}.',
       );
     }
 
-    final path = brick.path;
-    if (path != null) {
-      final brickYaml = File(p.join(path, BrickYaml.file));
-      if (!brickYaml.existsSync()) {
-        throw BrickNotFoundException(p.canonicalize(path));
-      }
-      final remoteDir = getPath(brick);
-      if (remoteDir != null) return remoteDir;
-      return _addLocalBrick(path);
-    }
-
-    return _addRemoteBrick(brick.git!);
-  }
-
-  /// Writes local brick at [path] to cache
-  /// and returns the local path to the brick.
-  String _addLocalBrick(String path) {
-    final localPath = p.canonicalize(path);
-    _cache[_getKey(Brick(path: path))!] = localPath;
+    final remoteDir = getPath(brick);
+    if (remoteDir != null) return remoteDir;
+    final localPath = p.canonicalize(brick.location.path!);
+    _cache[_getKey(brick)!] = localPath;
     return localPath;
   }
 
-  /// Writes remote brick at [gitPath] to cache
+  /// Writes remote brick via git url to cache
   /// and returns the local path to the brick.
-  Future<String> _addRemoteBrick(GitPath gitPath) async {
-    final key = _getKey(Brick(git: gitPath))!;
+  Future<String> _addRemoteBrickFromGit(Brick brick) async {
+    final gitPath = brick.location.git!;
+    final key = _getKey(brick)!;
     final dirName = _getKey(
-      Brick(git: GitPath(gitPath.url, ref: gitPath.ref)),
+      Brick.git(GitPath(gitPath.url, ref: gitPath.ref)),
     )!;
     final directory = Directory(p.join(rootDir.path, 'git', dirName));
     final directoryExists = directory.existsSync();
@@ -163,6 +214,17 @@ class BricksJson {
         if (directory.existsSync()) directory.deleteSync(recursive: true);
         throw BrickNotFoundException('${gitPath.url}/${gitPath.path}');
       }
+      final yaml = checkedYamlDecode(
+        brickYaml.readAsStringSync(),
+        (m) => BrickYaml.fromJson(m!),
+      );
+
+      if (yaml.name != brick.name) {
+        throw MasonYamlNameMismatch(
+          'Brick name "${brick.name}" '
+          'doesn\'t match provided name "${yaml.name}" in ${MasonYaml.file}.',
+        );
+      }
     }
 
     /// Even if a cached version exists, still try to update.
@@ -175,8 +237,11 @@ class BricksJson {
         await tempDirectory.rename(directory.path);
       } catch (_) {}
       _ensureRemoteBrickExists(directory, gitPath);
-      _cache[key] = directory.path;
-      return directory.path;
+      final localPath = p
+          .canonicalize(p.join(directory.path, gitPath.path))
+          .replaceAll(r'\', '/');
+      _cache[key] = localPath;
+      return localPath;
     }
 
     if (directoryExists) await directory.delete(recursive: true);
@@ -184,8 +249,11 @@ class BricksJson {
     await directory.create(recursive: true);
     await _clone(gitPath, directory);
     _ensureRemoteBrickExists(directory, gitPath);
-    _cache[key] = directory.path;
-    return directory.path;
+    final localPath = p
+        .canonicalize(p.join(directory.path, gitPath.path))
+        .replaceAll(r'\', '/');
+    _cache[key] = localPath;
+    return localPath;
   }
 
   Future<void> _clone(GitPath gitPath, Directory directory) async {
@@ -198,8 +266,122 @@ class BricksJson {
     }
   }
 
+  /// Writes remote brick from registry to cache
+  /// and returns the local path to the brick.
+  Future<String> _addRemoteBrickFromRegistry(Brick brick) async {
+    final version = await _resolveBrickVersion(brick);
+    final resolvedBrick = Brick.version(name: brick.name, version: version);
+    final key = _getKey(resolvedBrick)!;
+    final directory = Directory(p.join(rootDir.path, 'hosted', hostedUrl, key));
+    final directoryExists = directory.existsSync();
+    final directoryIsNotEmpty =
+        directoryExists && directory.listSync(recursive: true).isNotEmpty;
+
+    /// Use cached version if exists.
+    if (directoryExists && directoryIsNotEmpty) {
+      _cache[key] = directory.path;
+      return directory.path;
+    }
+
+    if (directoryExists) await directory.delete(recursive: true);
+
+    await directory.create(recursive: true);
+    await _download(resolvedBrick, directory);
+    _cache[key] = directory.path;
+    return directory.path;
+  }
+
+  Future<String> _resolveBrickVersion(Brick brick) async {
+    final _version = brick.location.version;
+    if (_version != null) {
+      try {
+        final version = Version.parse(_version);
+        return version.toString();
+      } catch (_) {}
+    }
+    final constraint = VersionConstraint.parse(_version ?? 'any');
+    final uri = Uri.parse('https://$hostedUrl/api/v1/bricks/${brick.name}');
+
+    late final http.Response response;
+    try {
+      response = await http.get(uri);
+    } catch (_) {
+      throw BrickResolveVersionException(
+        'Unable to fetch versions for brick "${brick.name}".',
+      );
+    }
+    if (response.statusCode == 404) {
+      throw BrickResolveVersionException(
+        'Brick "${brick.name}" does not exist.',
+      );
+    }
+
+    if (response.statusCode != 200) {
+      throw BrickResolveVersionException(
+        'Unable to fetch versions for brick "${brick.name}".',
+      );
+    }
+
+    late final Map body;
+    late final Version latestVersion;
+    try {
+      body = json.decode(response.body) as Map;
+      final latest = body['latest'] as Map;
+      final _latestVersion = latest['version'] as String;
+      latestVersion = Version.parse(_latestVersion);
+    } catch (_) {
+      throw BrickResolveVersionException(
+        'Unable to parse latest version of brick "${brick.name}".',
+      );
+    }
+
+    if (constraint.isAny) return latestVersion.toString();
+    if (constraint.allows(latestVersion)) return latestVersion.toString();
+
+    late final List<Version> versions;
+    try {
+      final _versions = body['versions'] as List;
+      versions = _versions
+          .map((dynamic v) => Version.parse((v as Map)['version'] as String))
+          .toList()
+        ..sort(Version.antiprioritize);
+    } catch (_) {
+      throw BrickResolveVersionException(
+        'Unable to parse available versions for brick "${brick.name}".',
+      );
+    }
+
+    for (final version in versions) {
+      if (constraint.allows(version)) return version.toString();
+    }
+
+    throw BrickUnsatisfiedVersionConstraint(
+      '"${brick.name}: $constraint" '
+      "doesn't match any versions.",
+    );
+  }
+
+  Future<void> _download(Brick brick, Directory directory) async {
+    final uri = Uri.parse(
+      'https://$hostedUrl/api/v1/bricks/${brick.name}/versions/${brick.location.version}.bundle',
+    );
+    final response = await http.get(uri);
+    if (response.statusCode != 200) {
+      throw BrickNotFoundException(uri.toString());
+    }
+
+    final bundle = MasonBundle.fromUniversalBundle(response.bodyBytes);
+    unpackBundle(bundle, directory);
+  }
+
   /// Global subdirectory within the mason cache.
   static Directory get globalDir => Directory(p.join(rootDir.path, 'global'));
+
+  /// The location of the registry where bricks are hosted.
+  static String get hostedUrl {
+    final environment = testEnvironment ?? Platform.environment;
+    return environment['MASON_HOSTED_URL'] ?? 'registry.brickhub.dev';
+  }
 
   /// Root mason cache directory
   static Directory get rootDir {