+++ packages/mason/lib/src/generator.dart
@@ -108,8 +108,25 @@ class MasonGenerator extends Generator {
   /// Factory which creates a [MasonGenerator] based on
   /// a [GitPath] for a remote [BrickYaml] file.
   static Future<MasonGenerator> fromGitPath(GitPath gitPath) async {
-    final directory = await BricksJson.temp().add(Brick(git: gitPath));
-    final file = File(p.join(directory, gitPath.path, BrickYaml.file));
+    final directory = await BricksJson.temp().add(Brick.git(gitPath));
+    final file = File(p.join(directory, BrickYaml.file));
+    final brickYaml = checkedYamlDecode(
+      file.readAsStringSync(),
+      (m) => BrickYaml.fromJson(m!),
+    ).copyWith(path: file.path);
+    return MasonGenerator.fromBrickYaml(brickYaml);
+  }
+
+  /// Factory which creates a [MasonGenerator] based on
+  /// a brick [name] and [version] hosted in a registry.
+  static Future<MasonGenerator> fromRegistry({
+    required String name,
+    required String version,
+  }) async {
+    final directory = await BricksJson.temp().add(
+      Brick.version(name: name, version: version),
+    );
+    final file = File(p.join(directory, BrickYaml.file));
     final brickYaml = checkedYamlDecode(
       file.readAsStringSync(),
       (m) => BrickYaml.fromJson(m!),