+++ packages/mason/test/src/exception_test.dart
@@ -1,6 +1,7 @@
 // ignore_for_file: prefer_const_constructors
 
 import 'package:mason/mason.dart';
+import 'package:mason/src/bricks_json.dart';
 import 'package:test/test.dart';
 
 void main() {
@@ -13,10 +14,18 @@ void main() {
       });
     });
 
-    group('WriteBrickException', () {
+    group('BrickResolveVersionException', () {
       test('can be instantiated', () {
         const message = 'test message';
-        final exception = WriteBrickException(message);
+        final exception = BrickResolveVersionException(message);
+        expect(exception.message, equals(message));
+      });
+    });
+
+    group('BrickUnsatisfiedVersionConstraint', () {
+      test('can be instantiated', () {
+        const message = 'test message';
+        final exception = BrickUnsatisfiedVersionConstraint(message);
         expect(exception.message, equals(message));
       });
     });
@@ -26,6 +35,13 @@ void main() {
         const path = 'test path';
         final exception = BrickNotFoundException(path);
         expect(exception.message, equals('Could not find brick at $path'));
+      });
+    });
+
+    group('MasonYamlNameMismatch', () {
+      test('has the correct message', () {
+        const message = 'test message';
+        expect(MasonYamlNameMismatch(message).message, equals(message));
       });
     });
   });