+++ packages/mason_cli/lib/src/command.dart
@@ -5,15 +5,6 @@ import 'package:mason/mason.dart';
 import 'package:path/path.dart' as p;
 import 'package:universal_io/io.dart';
 
-/// {@template mason_yaml_name_mismatch}
-/// Thrown when a brick's name in `mason.yaml` does not match
-/// the name in `brick.yaml`.
-/// {@endtemplate}
-class MasonYamlNameMismatch extends MasonException {
-  /// {@macro mason_yaml_name_mismatch}
-  MasonYamlNameMismatch(String message) : super(message);
-}
-
 /// {@template mason_yaml_not_found_exception}
 /// Thrown when a `mason.yaml` cannot be found locally.
 /// {@endtemplate}
@@ -91,9 +82,7 @@ abstract class MasonCommand extends Command<int> {
   /// Gets [BrickYaml] contents for bricks registered globally.
   Set<BrickYaml> get globalBricks {
     if (_globalBricks != null) return _globalBricks!;
-    return _globalBricks = _getBricks(
-      _getMasonYaml(_getMasonYamlFile(BricksJson.globalDir.path)),
-    );
+    return _globalBricks = _getBricks(globalBricksJson);
   }
 
   Set<BrickYaml>? _globalBricks;
@@ -102,7 +91,7 @@ abstract class MasonCommand extends Command<int> {
   Set<BrickYaml> get localBricks {
     if (_localBricks != null) return _localBricks!;
     return _localBricks = {
-      if (masonInitialized) ..._getBricks(masonYaml),
+      if (masonInitialized) ..._getBricks(localBricksJson!),
     };
   }
 
@@ -180,39 +169,19 @@ abstract class MasonCommand extends Command<int> {
   /// Return the current working directory.
   Directory get cwd => Directory.current;
 
-  /// The path to the cached brick directory if it exists.
-  /// Returns `null` if the brick is not cached.
-  String? _cacheDirectory(Brick brick) {
-    if (localBricksJson != null) {
-      final path = localBricksJson!.getPath(brick);
-      if (path != null) return path;
-    }
-    return globalBricksJson.getPath(brick);
-  }
-
   /// Gets all [BrickYaml] instances for the provided [masonYaml].
-  Set<BrickYaml> _getBricks(MasonYaml masonYaml) {
+  Set<BrickYaml> _getBricks(BricksJson bricksJson) {
     final bricks = <BrickYaml>{};
-    for (final entry in masonYaml.bricks.entries) {
-      final brick = entry.value;
-      final dirPath = _cacheDirectory(brick);
-      if (dirPath == null) break;
-      final filePath = brick.path != null
-          ? p.join(dirPath, BrickYaml.file)
-          : p.join(dirPath, brick.git?.path ?? '', BrickYaml.file);
+    for (final entry in bricksJson.cache.entries) {
+      final dirPath = entry.value;
+      final filePath = p.join(dirPath, BrickYaml.file);
       final file = File(filePath);
-      if (!file.existsSync()) throw BrickNotFoundException(filePath);
+      if (!file.existsSync()) throw BrickNotFoundException(dirPath);
       try {
         final brickYaml = checkedYamlDecode(
           file.readAsStringSync(),
           (m) => BrickYaml.fromJson(m!),
         ).copyWith(path: filePath);
-        if (brickYaml.name != entry.key) {
-          throw MasonYamlNameMismatch(
-            'brick name "${brickYaml.name}" '
-            'doesn\'t match provided name "${entry.key}" in ${MasonYaml.file}.',
-          );
-        }
         bricks.add(brickYaml);
       } on ParsedYamlException catch (e) {
         throw BrickYamlParseException(