+++ packages/mason_cli/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.1.0-dev.11
+
+- fix: clear `bricks.json` prior to fetching via `mason get`
+- fix: verify/validate brick name matches name in `mason.yaml` during `mason get`
+- refactor: populate bricks from `bricks.json` directly
+- refactor: remove dependency on `package:archive`
+
 # 0.1.0-dev.10
 
 - **BREAKING** feat: upgrade to `mason ^0.1.0-dev.6`